LayoutTests/svg/hittest/singular-transform-3.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
body { margin: 0; }
.p .a {
  transform: scaleX(0);
}
.p:hover .a {
  transform: scaleX(1);
  fill: red;
}
</style>
<svg width="400" height="100">
  <g class="p">
    <text class="a" y="80" font-size="100" font-family="Ahem">XX</text>
    <rect class="b" width="100" height="100" fill="green"/>
  </g>
</svg>
<script>
if (window.testRunner)
  eventSender.mouseMoveTo(150, 50);

shouldBeNull("document.querySelector('.p:hover .a')");
</script>
