third_party/blink/web_tests/svg/hittest/singular-transform-3-expected.txt

PASS document.querySelector('.p:hover .a') is null
PASS successfullyParsed is true

TEST COMPLETE

XX
